fog_gateway/requirements.txt
Flask>=2.0
pandas>=1.3
requests>=2.25
python-dotenv>=0.20
PyYAML>=6.0
